class Robot added

+++ Motor.cpp
@@ -7,13 +7,13 @@
 //#include "Util.h"
 
 extern int freeRam ();
-extern int16_t myAbs(int16_t param);
+extern int myAbs(int param);
 extern int sign(int val);
 
 //=========================================================
 void Motor::UpdateAccel()
 {  
-  int16_t absSpeed = abs( m_Speed );
+  int absSpeed = abs( m_Speed );
 
   if( absSpeed < SCURVE_LOW_SPEED )
   {
@@ -26,12 +26,12 @@ void Motor::UpdateAccel()
 }
 
 //=========================================================
-void Motor::UpdateSpeed( int16_t dt )
+void Motor::UpdateSpeed( int dt )
 {
   int tmp = m_Speed * m_Speed / ( 1800.0 * m_Accel );
   int stopPos = m_Step + sign(m_Speed) * tmp;
 
-  int16_t goalSpeed;
+  int goalSpeed;
   
   if( m_GoalStep > m_Step ) // Positive move
   {
@@ -47,12 +47,12 @@ void Motor::UpdateSpeed( int16_t dt )
 } // UpdateSpeed
 
 //=========================================================
-void Motor::SetSpeedInternal( int16_t dt, int16_t goalSpeed )
+void Motor::SetSpeedInternal( int dt, int goalSpeed )
 {
   goalSpeed = constrain( goalSpeed, -m_MaxSpeed, m_MaxSpeed );
   
   // We limit acceleration => speed ramp
-  int16_t accel = (long)m_Accel * dt * 0.001; // We divide by 1000 because dt are in microseconds
+  int accel = (long)m_Accel * dt * 0.001; // We divide by 1000 because dt are in microseconds
   if ( (long)goalSpeed - m_Speed > accel ) // We use long here to avoid overflow on the operation
   { 
     m_Speed += accel;
