+++ Motor.h
@@ -13,62 +13,62 @@ public:
   Motor();
   ~Motor();
   
-  void SetStep( int16_t step )
+  void SetStep( int step )
   {
   	m_Step = step;
   }
 
-  int16_t GetStep() const
+  int GetStep() const
   {
   	return m_Step;
   }
 
-  void SetGoalStep( int16_t step )
+  void SetGoalStep( int step )
   {
   	m_GoalStep = step;
   }
 
-  int16_t GetGoalStep() const
+  int GetGoalStep() const
   {
   	return m_GoalStep;
   }
 
-  void SetSpeed( int16_t speed )
+  void SetSpeed( int speed )
   {
     m_Speed = speed;
   }
 
-  int16_t GetSpeed() const
+  int GetSpeed() const
   {
     return m_Speed;
   }
 
-  void SetGoalSpeed( int16_t speed )
+  void SetGoalSpeed( int speed )
   {
     m_GoalSpeed = speed;
   }
 
-  int16_t GetGoalSpeed() const
+  int GetGoalSpeed() const
   {
     return m_GoalSpeed;
   }
   
-  void SetAccel( int16_t accel )
+  void SetAccel( int accel )
   {
     m_Accel = accel;
   }
 
-  int16_t GetAccel() const
+  int GetAccel() const
   {
     return m_Accel;
   }
   
-  void SetMaxAccel( int16_t accel )
+  void SetMaxAccel( int accel )
   {
     m_MaxAccel = accel;
   }
 
-  int16_t GetMaxAccel()
+  int GetMaxAccel()
   {
     return m_MaxAccel;
   }
@@ -89,16 +89,16 @@ public:
   }
   
   void UpdateAccel();
-  void UpdateSpeed( int16_t dt );
+  void UpdateSpeed( int dt );
   
 private:  
-  void SetSpeedInternal( int16_t dt, int16_t goalSpeed );
+  void SetSpeedInternal( int dt, int goalSpeed );
 
   //////////////
   // Position
   //////////////
-	int16_t m_Step;       // in motor steps. corresponds to position_M1/2
-  int16_t m_GoalStep;   // corresponds to target_position_M1/2
+	int m_Step;       // in motor steps. corresponds to position_M1/2
+  int m_GoalStep;   // corresponds to target_position_M1/2
 
   //////////////
   // Direction
@@ -108,12 +108,12 @@ private:
   ///////////////////
   // Speed & Accel
   ///////////////////
-	int16_t m_Accel;      // acceleration. corresponds to acceleration_M1/2
-  int16_t m_Speed;      // corresponds to speed_M1/2
-  int16_t m_GoalSpeed;  // corresponds to target_speed_M1/2
+	int m_Accel;      // acceleration. corresponds to acceleration_M1/2
+  int m_Speed;      // corresponds to speed_M1/2
+  int m_GoalSpeed;  // corresponds to target_speed_M1/2
 
-  int16_t m_MaxSpeed;   // max speed
-  int16_t m_MaxAccel;   // max acceleration
+  int m_MaxSpeed;   // max speed
+  int m_MaxAccel;   // max acceleration
   long    m_Period;     // for setting timer use. corresponds to timer_period. arduino "long" is 32 bit
   
 };
